backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
python-dotenv==1.0.0
langgraph==0.2.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
langchain-groq==0.2.0
tavily-python==0.3.0